requirements.txt
python-dotenv
tavily-python
google-genai
python-telegram-bot
yfinancefeedparser
